Consumer/requirements.txt
influxdb==5.3.1
influxdb-client==1.24.0
jsonschema==4.3.1
kafka-python==2.0.2
numpy==1.21.4
requests==2.26.0

